requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
asttokens==3.0.1
async-timeout==5.0.1
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
backoff==2.2.1
billiard==4.2.4
blinker==1.9.0
boto3==1.42.30
botocore==1.42.30
cachetools==6.2.4
catboost==1.2.8
category_encoders==2.8.1
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
configobj==5.0.9
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.4
dagshub-annotation-converter==0.1.15
databricks-sdk==0.78.0
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.25.2
dvc==3.66.1
dvc-data==3.18.2
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
gql==4.0.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.3.0
gto==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==8.38.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
kombu==5.6.2
lightgbm==4.6.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
missingno==0.5.2
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.7.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pathvalidate==3.3.1
patsy==1.0.2
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.19.1
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.15.3
scmrepo==3.6.1
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skops==0.13.0
smmap==5.0.2
SQLAlchemy==2.0.45
sqlparse==0.5.5
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.6
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.4.0
tomlkit==0.14.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
treelib==1.8.0
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
voluptuous==0.16.0
waitress==3.0.2
wcwidth==0.2.14
Werkzeug==3.1.5
xgboost==3.1.3
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
